pyproject.toml
[project]
name = "funcnodes-microscopy"
version = "0.1.0"
description = "A package to perform advanced microscopy data analysis"
readme = "README.md"
classifiers = [
    "License :: OSI Approved :: GNU Affero General Public License v3 or later (AGPLv3+)",
]
requires-python = "==3.11.*"
dependencies = [
    "funcnodes-core>=0.3.9",
    "funcnodes>=0.5.6",
    "funcnodes-react-flow>=0.3.3",
    "requests",
    "tifffile>=2025.1.10",
    "funcnodes-pandas>=0.2.11",
    "funcnodes-images>=0.2.4",
    "funcnodes-files>=0.2.11",
    "funcnodes-opencv>=0.2.4",
    "funcnodes-keras>=0.1.14",
    "stardist>=0.9.1",
    "scikit-image>=0.25.1",
    "super-image>=0.1.7",
    "funcnodes-sklearn>=0.1.15",
    "huggingface-hub==0.20.2",
]
[[project.authors]]
name = "Kourosh Rezaei"
email = "kourosh.rezaei@linkdlab.de"

[dependency-groups]
dev = [
    "funcnodes-module>=0.2.8",
    "pre-commit>=4.1.0",
    "pytest>=8.3.4",
]

[build-system]
requires = ["setuptools>=42"]
build-backend = "setuptools.build_meta"

[project.license]
text = "AGPL-3.0"

[project.urls]
homepage = "https://github.com/linkdlab/funcnodes_microscopy"
source = "https://github.com/linkdlab/funcnodes_microscopy"
tracker = "https://github.com/linkdlab/funcnodes_microscopy/issues"
download = "https://pypi.org/project/funcnodes-microscopy/#files"

[project.entry-points."funcnodes.module"]
module = "funcnodes_microscopy"
shelf = "funcnodes_microscopy:NODE_SHELF"

[tool.setuptools.package-dir]
"" = "src"

[tool.setuptools.packages.find]
where = ["src"]

# [tool.poetry.group.dev.dependencies]
# pre-commit = "*"
# pytest = "*"
# funcnodes-module = "^0.1.20"
